python/pyproject.toml
[project]
name = "valkey-playground-python"
version = "0.1.0"
description = ""
authors = [
    {name = "otajisan",email = "mtaji@morningcode.io"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "valkey-glide (==1.1.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
